Cleanup Notification logic
* We are only sending notifications when battery is below threshold
  hence no need for different type of icons/status/summary.

* Merge notification function into main.go
